+++ src/poll/poll.gateway.ts
@@ -6,7 +6,7 @@ import {
   OnGatewayConnection,
 } from '@nestjs/websockets';
 import { Server, Socket } from 'socket.io';
-import { Inject } from '@nestjs/common';
+import { Inject, Logger } from '@nestjs/common';
 import Redis from 'ioredis';
 import { REDIS_CLIENT } from '../redis/redis-client';
 import { DATABASE_CONNECTION } from '../database/database-connection';
@@ -14,16 +14,24 @@ import { PrismaClient } from '../common/generated/prisma/client';
 import { InjectQueue } from '@nestjs/bullmq';
 import { Queue } from 'bullmq';
 
+interface CastVotePayload {
+  roomId: string;
+  pollId: string;
+  userId: string;
+  supportsBunk: boolean;
+}
+
 @WebSocketGateway({
   cors: { origin: '*' },
   namespace: 'polls',
 })
 export class PollingGateway implements OnGatewayConnection {
   @WebSocketServer() server!: Server;
+  private readonly logger = new Logger(PollingGateway.name);
 
   constructor(
     @Inject(REDIS_CLIENT) private readonly redis: Redis,
-    @Inject(DATABASE_CONNECTION) private prismaService: PrismaClient,
+    @Inject(DATABASE_CONNECTION) private readonly prisma: PrismaClient,
     @InjectQueue('poll-sync') private readonly pollSyncQueue: Queue,
   ) {}
 
@@ -31,24 +39,59 @@ export class PollingGateway implements OnGatewayConnection {
     const roomId = client.handshake.query.roomId as string;
     if (roomId) {
       await client.join(`room:${roomId}`);
-      console.log(`Socket ${client.id} joined channel room:${roomId}`);
+      this.logger.log(`client:${client.id} joined room:${roomId}`);
     }
   }
 
   @SubscribeMessage('castVote')
-  async handleCastVote(
-    @MessageBody()
-    payload: {
-      roomId: string;
-      pollId: string;
-      userId: string;
-      supportsBunk: boolean;
-    },
-  ) {
+  async handleCastVote(@MessageBody() payload: CastVotePayload) {
     const { roomId, pollId, userId, supportsBunk } = payload;
+
+    if (!roomId || !pollId || !userId) return;
+
     const redisHashKey = `poll:${pollId}:votes`;
 
+    const [poll, roomMember] = await Promise.all([
+      this.prisma.bunkPoll.findUnique({ where: { id: pollId } }),
+      this.prisma.roomMember.findUnique({
+        where: { roomId_userId: { roomId, userId } },
+      }),
+    ]);
+
+    if (!poll) {
+      return this.server
+        .to(`room:${roomId}`)
+        .emit('pollError', { pollId, message: 'Poll does not exist.' });
+    }
+
+    console.log('Poll expired: ', poll.expiresAt.getTime() < Date.now());
+    console.log('Member not verified: ', !roomMember || !roomMember.isApproved);
+
+    if (poll.isLocked || poll.expiresAt.getTime() < Date.now()) {
+      return this.server.to(`room:${roomId}`).emit('pollError', {
+        pollId,
+        message: 'Poll is either locked or expired.',
+      });
+    }
+
+    if (!roomMember || !roomMember.isApproved) {
+      return this.server.to(`room:${roomId}`).emit('pollError', {
+        pollId,
+        message:
+          'You are not an eligible member in this room to cast your vote.',
+      });
+    }
+
     await this.redis.hset(redisHashKey, userId, String(supportsBunk));
+    const secondsUntilExpiry = Math.ceil(
+      (poll.expiresAt.getTime() - Date.now()) / 1000,
+    );
+    const safetyBuffer = 3600;
+    await this.redis.expire(redisHashKey, secondsUntilExpiry + safetyBuffer);
+
+    const approvedMembersCount = await this.prisma.roomMember.count({
+      where: { roomId, isApproved: true },
+    });
 
     const rawVotes = await this.redis.hgetall(redisHashKey);
     const totalVotes = Object.keys(rawVotes).length;
@@ -56,19 +99,11 @@ export class PollingGateway implements OnGatewayConnection {
       (v) => v === 'true',
     ).length;
 
-    const room = await this.prismaService.room.findUnique({
-      where: { id: roomId },
-      include: { members: true, polls: true },
-    });
-
-    const roomMembersCount =
-      room?.members.filter((mem) => mem.isApproved).length ?? 0;
-
-    const thresholdPercentage =
-      room?.polls.filter((poll) => poll.id === pollId)[0].threshold ?? 50;
-
-    const currentPercentage = (classSkipCount / roomMembersCount) * 100;
-    const consensusReached = currentPercentage >= thresholdPercentage;
+    const currentPercentage =
+      approvedMembersCount > 0
+        ? (classSkipCount / approvedMembersCount) * 100
+        : 0;
+    const consensusReached = currentPercentage >= poll.threshold;
 
     this.server.to(`room:${roomId}`).emit('pollUpdated', {
       pollId,
@@ -78,17 +113,24 @@ export class PollingGateway implements OnGatewayConnection {
     });
 
     if (consensusReached) {
-      await this.prismaService.bunkPoll.update({
+      await this.prisma.bunkPoll.update({
         where: { id: pollId },
         data: { isLocked: true },
       });
 
-      await this.pollSyncQueue.add('flush-to-postgres', { pollId });
+      await this.redis.del(redisHashKey);
 
       this.server.to(`room:${roomId}`).emit('pollLocked', {
         pollId,
         message: 'The pack has spoken. Class is skipped today.',
       });
     }
+
+    await this.pollSyncQueue.add(
+      'flush-to-postgres',
+      { pollId, rawVotes },
+      { jobId: `sync-${pollId}`, removeOnComplete: true },
+    );
+    this.logger.log(`poll-sync job queued with jobId:sync-${pollId}`);
   }
 }